+++ lua/gitlinker.lua
@@ -1,6 +1,7 @@
 local logger = require("gitlinker.logger")
 local Linker = require("gitlinker.linker").Linker
 local highlight = require("gitlinker.highlight")
+local deprecation = require("gitlinker.deprecation")
 
 --- @alias Options table<any, any>
 --- @type Options
@@ -96,163 +97,22 @@ local function setup(opts)
   end
 
   -- logger.debug("|setup| Configs:%s", vim.inspect(Configs))
-end
 
---- @deprecated
---- @package
---- @param pattern_rules table[]
---- @param remote_url string
---- @return string?
-local function _map_remote_to_host_deprecated(pattern_rules, remote_url)
-  for i, group in ipairs(pattern_rules) do
-    for pattern, replace in pairs(group) do
-      -- logger.debug(
-      --   "[map_remote_to_host] map group[%d], pattern:'%s', replace:'%s'",
-      --   i,
-      --   pattern,
-      --   replace
-      -- )
-      if string.match(remote_url, pattern) then
-        local host_url = string.gsub(remote_url, pattern, replace)
-        -- logger.debug(
-        --   "[map_remote_to_host] map group[%d] matched, pattern:'%s', replace:'%s', remote_url:'%s' => host_url:'%s'",
-        --   i,
-        --   pattern,
-        --   replace,
-        --   remote_url,
-        --   host_url
-        -- )
-        return host_url
-      end
-    end
+  if Configs.pattern_rules ~= nil then
+    deprecation.notify(
+      "'pattern_rules' is deprecated! please migrate to latest configs."
+    )
   end
-
-  return nil
-end
-
---- @param pattern_rules table
---- @return boolean
-local function _has_deprecated_pattern_rules(pattern_rules)
-  if type(pattern_rules) ~= "table" then
-    return false
+  if Configs.override_rules ~= nil then
+    deprecation.notify(
+      "'override_rules' is deprecated! please migrate to latest configs."
+    )
   end
-  for _, group in ipairs(pattern_rules) do
-    if type(group) ~= "table" then
-      return false
-    end
-    for pattern, replace in pairs(group) do
-      if
-        type(pattern) == "string"
-        and string.len(pattern) > 0
-        and type(replace) == "string"
-        and string.len(replace) > 0
-      then
-        return true
-      end
-    end
+  if Configs.custom_rules ~= nil then
+    deprecation.notify(
+      "'custom_rules' is deprecated! please migrate to latest configs."
+    )
   end
-  return false
-end
-
---- @package
---- @param remote_url string
---- @return string?
-local function _map_remote_to_host(remote_url)
-  local custom_rules = Configs.custom_rules
-  if type(custom_rules) == "function" then
-    local result = custom_rules(remote_url)
-    if result then
-      return result
-    end
-  end
-
-  local pattern_rules = Configs.pattern_rules
-  if _has_deprecated_pattern_rules(pattern_rules) then
-    local function notify()
-      local function impl()
-        local msg = string.format(
-          "[gitlinker] warning! detect deprecated 'config.pattern_rules', please migrate to latest schema."
-        )
-        local chunks = { { msg, "WarningMsg" } }
-        vim.api.nvim_echo(chunks, false, {})
-      end
-
-      vim.schedule(impl)
-      vim.defer_fn(impl, 3000)
-    end
-    notify()
-
-    logger.debug("|_map_remote_to_host| use deprecated pattern rules schema")
-    local result = _map_remote_to_host_deprecated(pattern_rules, remote_url)
-    if result then
-      return result
-    end
-  end
-
-  -- logger.debug("|_map_remote_to_host| use new pattern rules schema")
-  pattern_rules = vim.list_extend(
-    vim.deepcopy(Configs.override_rules or {}),
-    vim.deepcopy(pattern_rules)
-  )
-  for i, rule in ipairs(pattern_rules) do
-    local pattern = rule[1]
-    local replace = rule[2]
-    -- logger.debug(
-    --   "[map_remote_to_host] map group[%d], pattern:'%s', replace:'%s'",
-    --   i,
-    --   pattern,
-    --   replace
-    -- )
-    if string.match(remote_url, pattern) then
-      local host_url = string.gsub(remote_url, pattern, replace)
-      -- logger.debug(
-      --   "[map_remote_to_host] map group[%d] matched, pattern:'%s', replace:'%s', remote_url:'%s' => host_url:'%s'",
-      --   i,
-      --   pattern,
-      --   replace,
-      --   remote_url,
-      --   host_url
-      -- )
-      return host_url
-    end
-  end
-
-  return nil
-end
-
---- @param host_url string
---- @param lk Linker
---- @param opts Options?
---- @return string
-local function _make_sharable_permalinks(host_url, lk, opts)
-  local url = string.format([[%s%s/%s]], host_url, lk.rev, lk.file)
-
-  local add_plain = type(opts) == "table"
-    and type(opts.add_plain_for_markdown) == "boolean"
-    and opts.add_plain_for_markdown
-  local endswith_md = type(url) == "string"
-    and string.len(url) >= 3
-    and url:sub(#url - 2, #url):lower() == ".md"
-  -- logger.debug(
-  --     "|_make_sharable_permalinks| url:%s, add plain:%s, url sub:%s, lower:%s, endswith '*.md':%s",
-  --     vim.inspect(url),
-  --     vim.inspect(add_plain),
-  --     vim.inspect(url:sub(#url - 2, #url)),
-  --     vim.inspect(url:sub(#url - 2, #url):lower()),
-  --     vim.inspect(url:sub(#url - 2, #url):lower() == ".md")
-  -- )
-  if add_plain and endswith_md then
-    url = url .. [[?plain=1]]
-  end
-
-  if type(lk.lstart) == "number" then
-    url = string.format([[%s#L%d]], url, lk.lstart)
-    if type(lk.lend) == "number" and lk.lend > lk.lstart then
-      url = string.format([[%s-L%d]], url, lk.lend)
-    end
-  end
-
-  return url
 end
 
 --- @param opts Options?
@@ -298,8 +158,6 @@ end
 local M = {
   setup = setup,
   link = link,
-  _make_sharable_permalinks = _make_sharable_permalinks,
-  _map_remote_to_host = _map_remote_to_host,
 }
 
 return M
